+++ CHANGELOG.rst
@@ -30,7 +30,7 @@ can check out the `wish list <https://github.com/mborsetti/webchanges/blob/main/
    Security, in case of vulnerabilities. [triggers a minor patch]
    Internals, for changes that don't affect users. [triggers a minor patch]
 
-Version 3.9rc1
+Version 3.9rc2
 ===================
 Unreleased